test: remove the JWT keys the exchange suite writes, and share the helper
tokenExchange.test.js wrote real signing keys into the test base path and
never removed them. That path is shared:
unitTests/utility/install/checkJWTTokensExist.test.js asserts those files
are ABSENT (its happy path expects accessSync to throw ENOENT), and mocha
runs every file in one process against one base path — so whichever ran
first decided whether the other passed.

It has been latent here. It surfaced on the stacked branch (#2174) when a
second file started writing the same keys, failing that suite on all
three Node versions; the same landmine was already sitting on this branch
waiting for a file-order change.

The fix is a testUtils.installTestJwtKeys() that returns a cleanup
function, so the next test needing signing keys gets the removal for free
rather than copying the setup and not the teardown. Verified by running
the exchange suite and checkJWTTokensExist together, which failed before
and passes now.
